refactor: Update restaurants screen to use SafeAreaView and ExpoStatusBar
This commit refactors the restaurants.screen.jsx file to import and use the SafeAreaView component from the styled.js file. It also imports and uses the ExpoStatusBar component from the expo-status-bar package in the App component. These changes improve the rendering of the screen and manage the status bar in the app more effectively.